requirements.txt
datetime
requests
python-dotenv
Flask
pymongo
requests
hachoir
beautifulsoup4
aioschedule
colorama
pika
aiohttp
websocket-client
rel
cryptography
oauthlib
instaloader
cryptography